pc_client/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pystray==0.19.5
Pillow>=11.0.0
